+++ bigml/constants.py
@@ -328,6 +328,11 @@ PROPORTIONAL = 1
 OUT_NEW_FIELDS = "output_fields"
 OUT_NEW_HEADERS = "output_headers"
 
+# input data allowed formats in batch predictions
+DATAFRAME = "dataframe"
+INTERNAL = "list_of_dicts"
+
+
 IMAGE_EXTENSIONS = ['png', 'jpg', 'jpeg', 'gif', 'tiff', 'tif', 'bmp',
                     'webp', 'cur', 'ico', 'pcx', 'psd', 'psb']